[SYCL][NFC] Split the file with host builtins implementations to 5 parts
This splitting improved the build time and eliminated the problem with
too big object file on Windows/Debug.
